climate-quest-backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.100.1
h11==0.14.0
idna==3.4
packaging==23.1
plotly==5.15.0
pydantic-core==2.4.0
pydantic==2.1.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tenacity==8.2.2
typing-extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
python-dotenv==1.0.0 